+++ api/src/main/java/org/allaymc/api/permission/Permission.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public final class Permission {
     private static final PermissionListener COMMAND_PERMISSION_LISTENER = (permissible, value) -> {
         if (permissible instanceof EntityPlayer player) {
-            player.requireResendingCommands();
+            player.sendCommands();
         }
     };