+++ api/src/main/java/org/allaymc/api/world/WorldViewer.java
@@ -3,7 +3,7 @@ package org.allaymc.api.world;
 import org.allaymc.api.block.action.BlockAction;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.blockentity.BlockEntity;
-import org.allaymc.api.debugshape.DebugShape;
+import org.allaymc.api.debugshape.DebugShapeViewer;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.action.EntityAction;
 import org.allaymc.api.entity.component.EntityContainerHolderComponent;
@@ -22,7 +22,6 @@ import org.joml.Vector3dc;
 import org.joml.Vector3ic;
 
 import java.util.Collection;
-import java.util.Set;
 import java.util.UUID;
 
 /**
@@ -30,7 +29,7 @@ import java.util.UUID;
  *
  * @author daoge_cmd
  */
-public interface WorldViewer {
+public interface WorldViewer extends DebugShapeViewer {
     /**
      * Views the entity passed.
      *
@@ -228,39 +227,6 @@ public interface WorldViewer {
      * @param weather the weather to view
      */
     void viewWeather(Weather weather);
-
-    /**
-     * View a debug shape, which will make a specific debug shape being
-     * displayed for this viewer. This method is also used to update a
-     * debug shape in viewer's side when the debug shape is changed.
-     *
-     * @param debugShape the debug shape to view
-     */
-    void viewDebugShape(DebugShape debugShape);
-
-    /**
-     * View multiple debug shapes. This method is similar to {@link #viewDebugShape(DebugShape)}
-     * but can view multiple debug shapes at once.
-     *
-     * @param debugShapes the debug shapes to view
-     */
-    void viewDebugShapes(Set<DebugShape> debugShapes);
-
-    /**
-     * Remove a debug shape, which will make a specific debug shape being
-     * removed from this viewer.
-     *
-     * @param debugShape the debug shape to remove.
-     */
-    void removeDebugShape(DebugShape debugShape);
-
-    /**
-     * Remove multiple debug shapes. This method is similar to {@link #removeDebugShape(DebugShape)}
-     * but can remove multiple debug shapes at once.
-     *
-     * @param debugShapes the debug shapes to remove.
-     */
-    void removeDebugShapes(Set<DebugShape> debugShapes);
 
     /**
      * Views the block entity passed.